refactor: move component and swagger startup into DI modules
