+++ apps/server-telemetry/src/app/modules/swagger/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2025.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+export { SwaggerModule } from './module.ts';
