+++ apps/server-telemetry/src/cli/commands/start.ts
@@ -5,19 +5,8 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { generateSwagger } from '@privateaim/server-http-kit';
-import type { Component } from '@privateaim/server-kit';
-import { QueueWorkerComponentCaller, useLogger } from '@privateaim/server-kit';
+import { useLogger } from '@privateaim/server-kit';
 import { defineCommand } from 'citty';
-import path from 'node:path';
-import process from 'node:process';
-import {
-    EventEventQueueRouterRouting,
-    EventTaskQueueRouterRouting,
-    LogEventQueueRouterRouting,
-    LogTaskQueueRouterRouting,
-} from '@privateaim/server-telemetry-kit';
-import { EventComponent, LogComponent } from '../../app/components/index.ts';
 import { createApplication } from '../../app/index.ts';
 import { useEnv } from '../../app/modules/config/index.ts';
 
@@ -28,34 +17,10 @@ export function defineCLIStartCommand() {
             const app = createApplication();
             await app.setup();
 
-            const components : Component<any>[] = [
-                new QueueWorkerComponentCaller(
-                    new EventComponent(),
-                    {
-                        consumeQueue: EventTaskQueueRouterRouting,
-                        publishQueue: EventEventQueueRouterRouting,
-                    },
-                ),
-                new QueueWorkerComponentCaller(
-                    new LogComponent(),
-                    {
-                        consumeQueue: LogTaskQueueRouterRouting,
-                        publishQueue: LogEventQueueRouterRouting,
-                    },
-                ),
-            ];
-
-            await generateSwagger({
-                authupURL: useEnv('authupURL'),
-                baseURL: useEnv('publicURL'),
-                controllerBasePath: path.join(process.cwd(), 'src', 'adapters', 'http', 'controllers'),
-            });
-
-            const promises = components.map((c) => c.start());
-            await Promise.all(promises);
-
             const logger = useLogger();
-            logger.debug('Application started successfully.');
+
+            logger.debug(`Environment: ${useEnv('env')}`);
+            logger.debug(`Public-URL: ${useEnv('publicURL')}`);
         },
     });
 }
